refactor: useWaiting Store state와 socket으로 분리
