+++ src/pages/GamePage/GameMatchingPage/AutoGameMatchingPage/Auto1vs1/index.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
 import { useUsersMe } from '@/api';
-import { useWaitingSocket } from '@/api/socket/useWaitingSocket';
-import { useWaitingStore } from '@/api/store/useWaitingStore';
+import { useWaitingSocketStore } from '@/api/store/useWaitingSocketStore';
+import { useWaitingStore } from '@/api/store/useWaitingStateStore';
 import { Flex } from '@/components/system';
 import { BackButton } from '@/components/ui';
 
@@ -15,7 +15,7 @@ export const Game1vs1MatchingPage = () => {
   const { data } = useUsersMe();
   const meId = data?.data?.id;
 
-  const { socket } = useWaitingSocket();
+  const { socket } = useWaitingSocketStore();
   const [searchParams] = useSearchParams();
 
   const { users, tournamentSize, setTournamentSize } = useWaitingStore();
